Bug fixes for libvgfont and hello_font example
vgft.c is the most important.  Handling of strings with embedded
newlines was borked.  It was getting the wrong width when trying to
calculate the graphics extent for these strings.  Fixed and cleaned up.

Other vgfont changes are just to replace uint32_t with int32_t in all
the x and y offset arguments for plotting text.  Confirmed that negative
offsets behave correctly, as expected since ultimately these get passed
into VGInt, which is typedef khronos_int32_t, which is typedef int32_t.

The hello_font patch simply makes the resource transparent before first
displaying it, so you don't get a screen flash when it comes on.