Removed limitation on chars per line drawn
When drawing lines of characters, hardcoded buffers are used to store
glyph_indices and adjustments.  Presumably these aren't dealt with
char-by-char for efficiency purposes.  However, the hardcoded buffer was
introducing an artificial limit on how many characters vgfont could
handle on a line without missing an assert.

I refactored to draw the line iteratively in blocks if necessary to
reuse the glyph_indices and adjustments buffers, so there is no longer a
limitation on how many characters a line can have.